requirements.txt
requests>=2.31.0
pydantic>=2.0.0
Pillow>=10.0.0
numpy<2.0.0
torch